Remove exhaustiveness checks from InlineFragments derive (#442)
#### Why are we making this change?

Up until the recent changes in #391, cynic made use of the names of the
variants on an `InlineFragments` derive to determine what `__typename`
to look for.  This wasn't a great approach - it's clumsy having to name
the variants after the types they contain, and it's error prone: the
inner type could easily be for a different GQL type and cynic would not
be able to error.

PR #391 changed that - the `__typename` now comes from the inner
`QueryFragment`.  But I didn't remove the exhaustiveness check, which
still expected to get the `__typename` name from the variant name and
used that to validate things.

#### What effects does this change have?

This PR removes the exhaustiveness check, as there's not current an
easy way to do this at compile time, and this is more in line with how
GraphQL tends to work.

It's possible we can ressurect exhaustiveness checks at compile time in
the future using const assertions _or_ by generating unit tests that do
the check.  But for now I'm just removing.