Dont recase schema types (#483)
#### Why are we making this change?

The `use_schema` macro generates marker types for everything in a schema.
Prior to this change it would re-case those types to match rust casing norms -
usually `PascalCase` for types and `camelCase` for functions/fields.  It's not
clear that this was worthwhile though - users don't usually have to interact
with the contents of the schema module, and re-casing like this opens up some
odd edge cases.

For example #477 -  where a federated schema contains two `UUID` types with
different casing.  While this is definitely a weird edge case (and probably
_not_ something you'd want to do ideally), cynic should really be able to
handle it.

#### What effects does this change have?

It updates the `use_schema` macro to not re-case everything in the schema.
Instead keeping the original graphql schema casing, which is both easier and
less likely to cause problems

Fixes #477