Tolerate backwards compatible schema changes.
GraphQLs type coercion rules allow for certain changes to be made to a
schema without being considered a breaking change.  Amongst other
things:

1. Required fields in input types can be made optional without breaking
   clients.
2. Optional fields in output types can be made required.
3. Input fields can be made into lists without breaking, as a single
   item coerces into a list with a single item.

Cynic did not support this very well though - the type checking was
fairly strict and mostly expected a 1:1 mapping between the schema and
the rust types provided.

This PR updates the type checking to closer follow the graphql rules.

Fixes #239