Move `cynic_codegen::type_validation` into a types module.
The `type_validation` module in cynic-codegen has got pretty big.  Some
of it was private so it's not too bad, but it's got to the point that
the type parsing stuff is actually useful outside of this module.

So, this PR moves the code into a `types` module with two submodules -
one for parsing one for validation.  This should make it easier to do
some of the stuff I'm currently attempting to do.