Don't extract input objects from arguments anymore. (#526)
The generator uses the new GraphQL-style syntax for arguments in the
queries it generates.  Theres no reason to extract input types from
arguments when doing this as they're not likely to be used as rust
types - you just write a GraphQL literal instead.

So, this commit stops the generator from including these types in its
output.