Allow users to get __typename from union fallbacks (#576)
#### Why are we making this change?

When a user writes an InlineFragments struct they can provide a fallback
variant.  At present this is just a unit variant, but the user might like to
know which variant they have forgotten to include in their fragments.

#### What effects does this change have?

This change allows users to create a fallback variant for union types that gets
the `__typename` in it's single String field.

For example, for a GraphQL type

```graphql
union UnionType = Variant1 | Variant2 | Variant3 | Variant4
```

The following enum

```rust
#[derive(cynic::InlineFragments, Debug)]
pub enum UnionType {
    Variant1(Variant1Inner),
    Variant2(Variant2Inner),
    #[cynic(fallback)]
    FallbackVariant(String),
}
```

will capture the name of the variant in the field of the `FallbackVariant`.

Co-authored-by: Graeme Coupar <grambo@grambo.me.uk>