Re-organise the output of use_schema! (#577)
#### Why are we making this change?

The output of `use_schema` puts out structs for each type and a module of
fields in that type.  Previously these were all output in the same location: so
a type `foo` would end up with a module `foo_fields` that contains it's fields.
However if there was also a `foo_fields` type in the schema this would clash.

#### What effects does this change have?

Refactored the field modules so they live under a `__fields` module.  I suppose
this will also break if someone has a `__fields` type.  But _hopefully_ they
don't and I guess I'll re-think if they do.