Remove Send bound from ReqwestExt futures on wasm
When compiled for wasm, reqwest futures are not send as they contain a
websys type that is not Send.  Cynic uses a `BoxFuture` type that
expects a Send future, so doesn't compile when compiled for wasm with
http-reqwest enabled.

This commit updates the code to not expect a `Send` future under these
conditions which allows the code to compile.

Fixes #596