Turn ::cynic:: to cynic:: in derives outputs
This enables writing crates in the following structure:
- Sdk: exports schema module, tools and a prelude for query building
  containing custom types and cynic itself
- Applicative code: declares the queries modules using cynic querygen,
  and doing use xxx_gql_sdk::prelude::* at the top of every queries
  module

where we don't need to add the cynic dependency to every crate that
makes queries, and keep it in sync with the SDK's version, as this way
we can just rely on the existence of cynic in the namespace.

This breaks scenarios where people have modules named cynic in scope at
their query level, but I it looks like allowing this is less useful than
allowing the above-described pattern.