Make the GraphQLResponse deserialization stricter.
Currently both `data` & `errors` are Options, so you can throw
completely invalid JSON at it and the parse will succed.  This makes it
difficult to debug issues where there's some non-graphql error JSON in
the response.

I've added a custom Deserialize impl that validates at least one of them
is present and errors if not