Add schema registration functionality (#668)
#### Why are we making this change?

Currently all the cynic macros require access to the schema SDL.  This is a bit
annoying, because it means you need to provide the path to the schema whenever
you call one of these macros.

I added the `schema_for_derives` attribute macro to help alleviate this pain,
but it's also quite annoying.  It forces you to put things in a module when you
otherwise might not have to _and_ it often obscures the true source of macro
errors that occur within the module.

#### What effects does this change have?

This PR adds:

1. A `register_schema` macro that allows you to name a schema and provide it's
   SDL.  This will be moved into a known location inside the current builds
   `OUT_DIR`, allowing the other macros in `cynic` to refer to the schema by
   its name instead of it's path.
2. The above also lets you specify a schema as a "default".  I intend to update
   the macros to use this schema when no alternative is provided to them.  This
   should allow the user to omit schema information entirely in the common case
   where they're only working with one remote schema.
3. Added a `schema` attribute macro, intended to replace `use_schema` when
   paired with registered schemas.  This can be used on a schema module and
   will pull in pre-generated schema output from `OUT_DIR`.

This functionality replaces the older `output_schema_module` functionality, 
which I have marked as deprecated.