Allow cynic-codegen to use a preprocessed schema (#660)
#### Why are we making this change?

Certain schemas (GitHub in particular but I'm sure there's others) are massive.
If you build a cynic project with these it can take a while to compile, and
rust-analyzer sometimes goes a bit unresponsive.

As far as I can tell _some_ of this is down to the astronomical quantity of
code `use_schema` puts out.  But each of the derives also has to parse the
entire schema - which is a lot of parses if you're doing anything more than the
simplest query.

`graphql_parser` is relatively fast - it can parse the GH schema in 9ms in a
release build and 50ms in a debug build.  But particularly for debug builds,
this adds up - even just 10 parses of the schema is half a second wasted on
repeat work.

#### What effects does this change have?

This PR adds `rkyv` into `cynic-codegen` and offers functionality to pre-parse
the schema into an rkyv format in your `build.rs`, which the derives can then
load rather than parsing the SDL every time.

In my initial testing (using triagebot as a test case) this shaves somewhere
between 30-50% off the compile time of the crate that uses cynic.  It does also
add compile time to build rkyv, but that should be an infrequent cost on
developers machines, so this should be an overall win.

Part of #150 (need to see if this + other optimisations are finally enough)