Expose pre-registered schema functionality on derives. (#670)
PRs #668 & #660 exposed functionality for registering schemas in your
build.rs

This PR updates the various derive macros to accept the schema name
instead of the path, which vastly simplifies the API for these. They
also support _not_ specifying a schema, in which case they'll attempt to
use the one that was registered as the default (or error if no default
was registered).

I've not updated the schema_for_derives attribute macro as part of this,
as I believe it might be more trouble than its worth with this new
approach. I'll try it out and see - either update that macro or possibly
deprecate it if it's no longer any use.

Fixes #669