Improve errors for unregistered schemas.
I want to make registering schemas the recommended behaviour, and update
querygen accordingly.  But if people don't know what they're doing they'll
easily run into the various error cases for unregistered schemas.

This commit adds a ui-test for that case and improves the errors that appear
for that test, as they were not really clear _at all_.