Add fallback support to enum derive (#692)
#### Why are we making this change?

Some servers will add new enum variants without considering this a
breaking change. In this scenario we want to provide users with a way to
keep their cynic clients working.
  
#### What effects does this change have?

Adds support for the `#[cynic(fallback)]` attribute on a single enum
variant, similar to the one in `InlineFragments`. When this is present
it will be used for any enum variants we don't understand. It can
optionally have a single `String` field that will get the name of the
variant.

Fixes #629