Adds Operation::new (#652)
Provides a mechanism for constructing Operations for pre-determined
queries outside of this crate. Useful for test code when, for example,
users are writing their own HTTP integration.

#### Why are we making this change?

I'm interested in using `cynic`, but need to write a custom HTTP
integration. (My company have their own internal HTTP crate). To write
unit tests for this integration, I need to construct instances of
`Operation` but don't need to use cynic's builders.

It's not currently possible to construct instances of `Operation`
outside of the crate, because the `phantom` field is private.

#### What effects does this change have?

This is an additive, non-breaking change.

---------

Co-authored-by: Louis Rose <louis.rose@anaplan.com>
Co-authored-by: Graeme Coupar <grambo@grambo.me.uk>