Add 2021 introspection query into cynic (#727)
#### Why are we making this change?

v3.0.0 added `cynic-introspection` for running introspection queries
against a GraphQL server. It used a query that's compatible with servers
implementing the 2018 GraphQL spec. In the 2021 GraphQL spec the
`specifiedBy` field was added to `__Type` to support the `@specifiedBy`
directive on scalar definitions.

#### What effects does this change have?

This updates `cynic-introspection` with support for querying
`specifiedBy` in a new `query2021` module. This is intended to be
backwards compatible - users using `query` will continue to make 2018
queries, and only if they opt in to 2021 will they run the new query.

The `Schema` abstraction supports both queries and should probably be
what users use rather than using the queries directly.

Fixes #686