Added CapabilityDetectionQuery (#731)
#### Why are we making this change?

I've been working on adding introspection support to cynic recently.  

The october 2021 version of the specification added new fields into
introspection, to cover `@specifiedByUrl`. This presents a challenge to
introspection queries - if you request this field from a server that
doesn't support it you'll get an error. This is fine if you know which
version of the spec you're targeting, or you don't care about this
field. But if you are querying arbitrary servers it's a problem.

#### What effects does this change have?

This adds a `CapabilityDetectionQuery` into `cynic-introspection` - a
very cut down introspection query and some associated logic that can be
used to determine what a server supports. Currently this just detects
which version of the specification is supported, but I intend to extend
it to detect `@oneOf`, `@stream`, `@defer` and maybe others as and when
the need arises.