Make cynic-introspection capability aware (#738)
#### Why are we making this change?

#686 added support for the 2021 introspection query by adding some new
structs. This never felt like the best solution: I'd also like cynic to
support servers that implement `@oneOf` or other RFCs, and adding a new
set of structs for each combination of versions & features is not going
to scale well.

#### What effects does this change have?

Updates `cynic-introspection` to use the feature flags feature
introduced in #736. This means that we have one introspection query that
can support all the permutations of the introspection schema that a
remote might support.

Currently this just supports the two versions of the specification, but
it could easily be extended to support the various RFCs