Add cargo-dist  (#740)
#### Why are we making this change?

Cynic now has a CLI. Although users can probably install this via `cargo
install`, ideally we need some way to upload binaries to GitHub on
release. This will probably be better for CI etc.

#### What effects does this change have?

Adds `cargo-dist` into the repo so it can upload files on release. I've
not enabled the installer building code just yet, but we could do that
later on if it seems like a good idea.