Support QueryVariables with unused variables
GraphQL servers will error if you send a query with a declared variable that
isn't used.  If a user of cynic wants to use the same QueryVariables struct on
two queries then this means they need to make sure to use all the variables, which isn't ideal.

This is also a case of something I try to avoid in cynic: a query that compiles yet will never work on the remote server.

So, I've fixed it by detecting which variables are used from a given struct and only declaring those in the query string.

Fixes #768