First pass of cynic-parser (#803)
#### Why are we making this change?

There are parts of the latest spec that `graphql-parser` doesn't
support, so I want to replace it with an alternative.

Other options out there that I'm aware of:

- `async-graphql-parser` would work but its AST is quite annoying to
work with, and last time I tested it was slower than `graphql-parser`.
- `apollo-parser` continues on failure to parse, which is nice but means
there are a lot of `Option`s in its AST. I don't particularly want that
behaviour so that's out.
- `apollo-compiler` would probably make `apollo-parser` workable, but
feels like too heavy a dependency for my needs.

So it looks like I might write my own. Also, it'll be fun.

#### What effects does this change have?

Adds the first pass of a parser for GraphQL type system documents.