Handle disabled introspection better (#840)
#### Why are we making this change?

It's common for GraphQL servers to disable introspection in production.
Sometimes this is implemented by returning `null` for the `__schema`
field. Technically this breaks the schema for introspection, which
states that type is non-null, but people do it so what can I do.

#### What effects does this change have?

Updates the introspection query used by cynic-introspect so that decodes
still succeed in the face of this. Users of `cynic-introspection` can
then look in the errors to spot the "Introspection disabled" message
that is no-doubt hiding within.

Also updated `cynic-cli` to handle this better