Implement parsing of executable documents (#850)
#### Why are we making this change?

I've been working on adding a parser to cynic-parser recently. But up
till now it's had no support for executable documents.

#### What effects does this change have?

Adds the bulk of the AST & parsing code for parsing executable
documents. This isn't anywhere near finished yet, but since the AST and
grammar are mostly there it seemed a sensible point for a PR.