generate executable ast in cynic-parser (#870)
#### Why are we making this change?

The cynic-parser AST has a lot of fairly boilerplatey code in it. Each
type we store needs a record, a reader, some mechanical functions on the
reader that hook up to the underlying data and a few other bits &
pieces.

Maintaining all this by hand is going to be a lot of work, particularly
if I want to make sweeping changes (which I do in the type_system - I
want to re-arrange it completely).

#### What effects does this change have?

This adds an ast-generator crate that uses the parser to parse a type
system that defines the executable AST. We then use this to generate all
the records & readers for that AST. Also acts as a nice dogfood of the
parser.