fix: use github app creds for cli git in release (#891)
#### Why are we making this change?

I noticed that CI wasn't running on the release PR #887. This seems to
be because release-plz does force pushes with the git CLI instead of via
the API.

#### What effects does this change have?

Attempts to configure the git CLI to use the same GitHub app token that
release-plz uses for the API, which should hopefully fix this.