feat: add iterator type to cynic-parser (#945)
I realised while working on Debug impls that having every function in
cynic-parser return an anonymous `Iterator` isn't ideal. It's fine, but
there's a lot of different `Iterator` traits, and specifying each of
them in the `impl` return is laborious.

In particular I wanted to add a `Debug` impl for each of the `Iterators`
so we didn't need to collect just to do a `Debug`.

This adds an `Iter` trait to each of the modules in `cynic-parser`. They
each implement a whole host of `Iterator` traits, and also `impl Debug`
so we no longer need to collect in our Reader Debug impls.