fix: a bunch of formatting inconsistencies (#957)
#### Why are we making this change?

My goal with the formatting in cynic-parser is to be compatible with
prettier. If a user dumps a schema from cynic-parser, and then runs
prettier, that file should be unchanged. This should reduce format churn
and make it easier to use both prettier & cynic-parser at the same time.

But it's not quite there yet.

#### What effects does this change have?

Fixes a few issues I found in the formatting