feat: expose ids in cynic-parser (#959)
#### Why are we making this change?

The readers in cynic-parser all just store an id that indexes into the
underlying parsed Document. I was a bit on the fence about whether to
exppose these ids or not. But I've run into a use case for the
executable AST that'd be hard to do without the ids.

#### What effects does this change have?

1. Exposes an `id` function on all of the object readers.
2. Unions are unfortunately a bit more tricky, because they've generally
throw away their ID in favour of an enum with readers inside that. So
for this case, I've added an `ids` function to the `Iter` type, which I
think covers most of those cases. Can revisit if it doesn't.