move pretty printing under pretty feature flag (#960)
#### Why are we making this change?

cynic-parser currently only has a pretty printer. This is nice, but
there are some situations where that's not suitable. Particularly if you
care about performance over correctness.

#### What effects does this change have?

I'd like to add simpler printing based on Display impls, and use the
`print` feature flag to do that. This PR is prep for that, moving the
pretty printing under a new feature flag `pretty`

BREAKING CHANGE: this is definitely a breaking change