fix: fragments on interfaces in the generator (#994)
The generator contains a function `field_from_path` that accepts a root
type and a path. This was originally written before I supported
fragments, so assumed that the root type would always be an object (as
it is for query, mutation & subscription roots). For fragments though
the root type can also be an interface or a union type.

This partially fixes the problem so that we support interfaces. I didn't
add support for unions because unions don't specifically have fields
(except `__typename`) and I can't remember how this stuff works well
enough to fix it for unions right now.

Suspect there might be other problems related to this in there, but
that's a problem for another day.

Fixes #993