feat: add QueryVariableLiterals (#1009)
#### Why are we making this change?

In #1000 & #1004 a user is wanting to build a query that doesn't make
use of variables without hardcoding the values of arguments. This is to
make use of [some shopify functionality][1] that requires you to submit
a query without variables.

My initial instinct was that this should be done externally to cynic,
but I did plan on documenting how. #1007 was added to cater to this - it
contained an example that used the `graphql-query` crate to post-process
a cynic query and inline all of the variables into arguments in the
query.

However, when writing this I discovered it's not quite as smooth as I'd
hoped: you can't just dump the variables into JSON and put those into
the query because enums are serialised as strings in JSON but enum
format in GraphQL. Being unable to use JSON means you're also unable to
dynamically look up variables from their names - you have to hardcode
every variable that you want to substitute. Which is pretty crap.

#### What effects does this change have?

This PR introduces a new `QueryVariableLiterals` trait & derive that
aims to help with this: it exposes a single function `get` that can be
passed the name of a variable and it will return the appropriate
`InputLiteral` for that variable.

This can be used to make a `graphql-query` based transform much nicer,
but it should also make it easy enough to add support for variable
inlining into cynic itself.

[1]: https://shopify.dev/docs/api/admin-graphql/2024-07/mutations/bulkoperationrunquery