fix!: support ints larger than i32 in parser (#1037)
The GraphQl `Int` scalar is an i32. In cynic-parser I made the parsed
`Int` an i32 as well. However on closer inspection the spec doesn't
really impose the i32 limit on `IntValue`s as represented in the
language.

It's possible that a custom `BigInt` scalar or similar might want to
represent itself as a literal integer > an i32, and currently this is
unsupported in cynic-parser.

This change adds a new `IntValue` wrapper around ints in the `Value`
enums. This hides the internal representation a bit, allowing us to
change it later. Also added support for numbers up to i64.