feat: sorting in cynic-parser pretty printing (#1038)
#### Why are we making this change?

I'm writing some tests that compare the output of the same schema but
from two different sources. I'm using cynic-parsers pretty printing to
get consistent formatting of these two schemas. But the order that the
definitions are output in between the two schemas does not match.

#### What effects does this change have?

Adds an optional setting to the cynic-parser formatting that causes the
definitions & fields of the schema to be sorted prior to output.