feat: refactor IdReader APIs in parser (#1064)
#### Why are we making this change?

The `IdReader` API used to rely on the implementation of
`TypeSystemId`/`ExecutableApi` to create Readers. This meant that
working with this in a generic way had some quite complicated trait
bounds, and also means that an `Id` can only be associated with one kind
of `Reader`.

#### What effects does this change have?

Adds a `new` function into the `IdReader` api which simplifies the trait
bounds and also allows multiple `Readers` to use the same `Id` should
they want to.