feat: add `Type::definitions` (#1067)
#### Why are we making this change?

If you're working with a TypeSystemDocument, it's often useful to be
able to go from a `Type` to the definition(s) of that type. Currently
this is a bit awkward: you need to pass the `TypeSystemDocument` around,
potentially through several layers of indirection, so that you can use
it again to look up the type.

#### What effects does this change have?

Adds a `Type::definitions` function that returns an `Iterator` of
definitions that have the name contained within the `Type`. This is
currently not very efficient, but could be made more so later by adding
indexes.