fix: remove lifetimes from IdReader trait (#1072)
https://github.com/obmarg/cynic/pull/1064 improved some aspects of the
IdReader API, but by using a lifetime parameter on the `IdReader` trait
to implement `Iter` we tie the lifetimes in `Iter<'_, Whatever<'a>>`
together. This forces any usage sites of `Iter` to always have named
lifetime parameters - otherwise the compiler complains.

This is not the API I want to expose: the dual lifetimes are only
workable as they can be elided.

This fixes that somewhat. It does regress some of the problems #1064 was
meant to fix, but it's an acceptable tradeoff.