chore: remove reliance on hosted swapi (#1119)
The hosted swapi GraphQL endpoint doesn't seem to be working at the
moment, and it's causing all the tests to fail. Not ideal.

This adds a graphql-mocks crate (ported & simplified from an
implementation in the grafbase repo - probably not the cleanest since I
did the port quick n dirty, but it works) and updates the tests to use
that.