docs: import derive rust docs from book
For ages the book has been the canonical source of documentation for
cynic.  However, it's a bit annoying having the rust documentation
completely empty and just linked to the book.  Particularly since every
version of the crate links to just a single version of the rust docs.

This is an attempt to somewhat fix that: instead of just linking to the
book, we'll pull in the docs using the (relatively) new `include_str!`
support inside attribute macros.

This will require a bit of a reworking of the book, since there's a lot
of relative links to the book (and also parts of the book that won't
make sense as rust docs).  But it seems like a good idea generally.