feat: field type overrides in generator (#1184)
#### Why are we making this change?

I'm using `cynic-querygen` for a work project, and it's been a joy to
simply copy queries from Graphiql to our rust codebase, so thank you for
that! However, we've ended up needing to make quite a few
find-and-replace changes to the generated code due to the crate's
limited code generation options.

We'd like to start contributing some of the features we've found we
needed as new options to `cynic-querygen`. This first one is the one
that's the hardest to do via regex, because it should incorporate both
the type name and the field name, and also seemed like one of the
simplest to implement.

The main use case for this feature for us is specifying different
deserialization targets for different `Json` fields. For instance, we
have one JSON field that should deserialize to `HashMap<String,
String>`, and another that should deserialize directly to a complex enum
structure.

#### What effects does this change have?

This adds a new `field_overrides` field to `QueryGenOptions`, which
allows overriding the type of individual fields on specific types using
a `.` separator. I have no strong opinion on the shape of the option,
this just seemed the most natural. This field is passed down into
`make_query_fragment`, where each field checks it to potentially set its
`type_name_override`. For example:

```rust
let result = cynic_querygen::document_to_fragment_structs(
    &query_content,
    &schema_content,
    &cynic_querygen::QueryGenOptions {
        schema_name: Some("my_schema".to_string()),
        field_overrides: maplit::hashmap!(
            "MyType.complexEnum" => "crate::complex_enum::ComplexEnum",
            "PresignedUrl.fields" => "std::collections::HashMap<String, String>",
        ),
        ..Default::default()
    },
);
```

Test and documentation are included for the new field, and I also added
a doc comment for `schema_name`, since I was surprised when I originally
looked into `cynic-querygen` to find it basically lacking any
documentation.

If you like the idea of making the querygen more customizable, we have a
couple other ideas I'd like to follow up with, which would fully
supplant all our find-and-replaces:

- overriding the default type for scalars like `DateTime`, `Json`, and
`Uuid` (e.g. to use `chrono`, `serde_json` and `uuid` crate types)
- specifying additional traits to derive for `cynic::QueryFragment`s and
`cynic::Enum`s